IGNITE-3670 IGFS: Improved symlink handling for delete operation and added more tests. This closes #975.
